# Add sample statistics to operator telemetry