--- a/execution/model/operator.go
+++ b/execution/model/operator.go
@@ -9,12 +9,16 @@ import (
 	"time"
 
 	"github.com/prometheus/prometheus/model/labels"
+	"github.com/prometheus/prometheus/util/stats"
 )
 
 type OperatorTelemetry interface {
 	AddExecutionTimeTaken(time.Duration)
 	ExecutionTimeTaken() time.Duration
 	fmt.Stringer
+	IncrementSamplesAtStep(samples int, step int)
+	UpdatePeak(samples int)
+	Samples() *stats.QuerySamples
 }
 
 func NewTelemetry(operator fmt.Stringer, enabled bool) OperatorTelemetry {
@@ -39,10 +43,18 @@ func (tm *NoopTelemetry) ExecutionTimeTaken() time.Duration {
 	return time.Duration(0)
 }
 
+func (tm *NoopTelemetry) IncrementSamplesAtStep(_, _ int) {}
+
+func (tm *NoopTelemetry) UpdatePeak(_ int) {}
+
+func (tm *NoopTelemetry) Samples() *stats.QuerySamples { return stats.NewQuerySamples(false) }
+
 type TrackedTelemetry struct {
 	name          string
 	ExecutionTime time.Duration
 	fmt.Stringer
+
+	LoadedSamples *stats.QuerySamples
 }
 
 func NewTrackedTelemetry(operator fmt.Stringer) *TrackedTelemetry {
@@ -58,6 +70,16 @@ func (ti *TrackedTelemetry) AddExecutionTimeTaken(t time.Duration) { ti.Executio
 func (ti *TrackedTelemetry) ExecutionTimeTaken() time.Duration {
 	return ti.ExecutionTime
 }
+
+func (ti *TrackedTelemetry) IncrementSamplesAtStep(samples, step int) {
+	ti.LoadedSamples.IncrementSamplesAtStep(step, int64(samples))
+}
+
+func (ti *TrackedTelemetry) UpdatePeak(samples int) {
+	ti.LoadedSamples.UpdatePeak(samples)
+}
+
+func (ti *TrackedTelemetry) Samples() *stats.QuerySamples { return ti.LoadedSamples }
 
 type ObservableVectorOperator interface {
 	VectorOperator