--- a/storage/prometheus/matrix_selector.go
+++ b/storage/prometheus/matrix_selector.go
@@ -161,10 +161,12 @@ func (o *matrixSelector) Next(ctx context.Context) ([]model.StepVector, error) {
 		ts += o.step
 	}
 
+	stepSamples := int64(0)
 	// Reset the current timestamp.
 	ts = o.currentStep
 	firstSeries := o.currentSeries
 	for ; o.currentSeries-firstSeries < o.seriesBatchSize && o.currentSeries < int64(len(o.scanners)); o.currentSeries++ {
+		stepSamples = 0
 		var (
 			series   = &o.scanners[o.currentSeries]
 			seriesTs = ts
@@ -205,7 +207,10 @@ func (o *matrixSelector) Next(ctx context.Context) ([]model.StepVector, error) {
 				} else {
 					vectors[currStep].AppendSample(o.vectorPool, series.signature, f)
 				}
+				stepSamples += int64(len(series.buffer.Samples()))
 			}
+			o.UpdatePeak(int(stepSamples))
+			o.IncrementSamplesAtStep(int(stepSamples), currStep)
 			seriesTs += o.step
 		}
 	}