@pedro-stanaka 2024-03-22
> Hm, do we want to track processed sample for each operator, or only for scanners?

I thought of adding to operators because in some cases you lose context of usage, like in Execution operator.