@fpetkovski 2024-03-24
Maybe we can run benchmarks to see if counting each sample individually is going to have a significant perf penalty.