@pedro-stanaka 2024-03-25
> Maybe we can run benchmarks to see if counting each sample individually is going to have a significant perf penalty.

The change is pretty minimal, you can check the benchmarks on the PR description.